Add files via upload
This code is for a Risk detection model that employs a Multi-LLM approach utilizing three large language models (LLMs). The model aggregates the opinions of the three LLMs to make the final risk detection decision. The dataset used is the Korean public dataset, and the model analyzes both past and future contexts before each of the three LLMs makes its final judgment.